Add Suunto JSON import and modularize import system
- Implement Suunto JSON parser supporting standard and Ocean formats
- Extract CNS, OTU, and SAC rate from Suunto exports
- Refactor monolithic dives_import.py into modular imports/ sub-package
- Modularize logic for Subsurface XML, Garmin FIT, and SSI CSV
- Fix circular dependency by creating gas_utils for shared logic
- Update all existing tests to use new modular import structure
- Harden security by preventing information exposure in all import endpoints
- Update test script to make coverage reporting optional via --cov flag
- Add Suunto import support to frontend modal and services
