+++ backend/app/routers/dives/imports/__init__.py
@@ -1,0 +1,5 @@
+from . import subsurface
+from . import garmin
+from . import ssi_csv
+from . import suunto
+from . import confirm
