+++ backend/pytest.ini
@@ -8,9 +8,6 @@ addopts =
     --tb=short
     --strict-markers
     --disable-warnings
-    --cov=app
-    --cov-report=term-missing
-    --cov-report=html
     --asyncio-mode=auto
 markers =
     auth: Authentication tests
